// src/api/models/Contact.ts
export interface ContactPayload {
  firstName: string;
  lastName: string;
  birthdate?: string;
  email?: string;
  phone?: string;
  street1?: string;
  street2?: string;
  city?: string;
  stateProvince?: string;
  postalCode?: string;
  country?: string;
}

export interface ContactResponse extends ContactPayload {
  _id: string;
  owner: string;
}